backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.6
pydantic-settings>=2.2
supabase>=2.4
google-generativeai>=0.5
python-dotenv>=1.0
python-multipart>=0.0.9
sse-starlette>=2.0
